Changes
-added interface
-divide repository class to sub classes
