+++ .idea/libraries/postgresql_42_7_1.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="postgresql-42.7.1">
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/assignment 3/src/postgresql-42.7.1.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
