+++ .idea/libraries/postgresql_42_7_12.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="postgresql-42.7.12">
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/Database connection/lib/postgresql-42.7.1.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
