+++ assignment 3/src/controllers/MovieController.java
@@ -1,14 +1,14 @@
 package controllers;
 
 import entities.Movie;
-import repositories.Repositories;
+import repositories.MovieTable;
 
 import java.sql.SQLException;
 import java.util.Scanner;
 
 public class MovieController {
-    private Repositories movieRepository;
-    public MovieController(Repositories userRepository) {
+    private MovieTable movieRepository;
+    public MovieController(MovieTable userRepository) {
         this.movieRepository = userRepository;
     }
 
@@ -17,33 +17,33 @@ public class MovieController {
             String movieName = scanner.nextLine();
             System.out.print("Enter price: ");
             double price = Double.parseDouble(scanner.nextLine());
-            movieRepository.addMovie(new Movie(movieName, price));
+            movieRepository.addToTable(new Movie(movieName, price));
     }
     public void updateMovie(Scanner scanner) throws SQLException{
             System.out.println("Enter the id to update");
             int id = Integer.parseInt(scanner.nextLine());
-            if(movieRepository.getMovie(id)) {
+            if(movieRepository.getObjectFromTable(id)) {
                 System.out.println("Enter the movie name to update: ");
                 String movieName = scanner.nextLine();
 
                 System.out.println("Enter the price to update: ");
                 double price  = Double.parseDouble(scanner.nextLine());
 
-                movieRepository.updateMovie(new Movie(id, movieName, price ));
+                movieRepository.updateObjectTable(new Movie(id, movieName, price ));
             }
     }
     public void deleteMovie(Scanner scanner) throws SQLException{
                 System.out.println("Enter the id to delete");
                 int id = Integer.parseInt(scanner.nextLine());
-                if(movieRepository.getMovie(id)) {
-                    movieRepository.deleteMovie(id);
+                if(movieRepository.getObjectFromTable(id)) {
+                    movieRepository.deleteObjectFromTable(id);
                 }
     }
      public void getMovie(Scanner scanner) throws SQLException{
         Scanner sc = new Scanner(System.in);
         System.out.println("Enter id of movie");
         int id = sc.nextInt();
-        movieRepository.getMovie(id);
+        movieRepository.getObjectFromTable(id);
      }
      public void getMovieByName(Scanner scanner) throws SQLException{
         Scanner sc = new Scanner(System.in);
@@ -54,7 +54,7 @@ public class MovieController {
 
 
     public void getAllMovies() throws SQLException{
-        movieRepository.getAllMovies();
+        movieRepository.getAllObjectsFromTable();
     }
 
 }
