+++ assignment 3/src/controllers/UserController.java
@@ -1,14 +1,15 @@
 package controllers;
 import entities.User;
-import repositories.Repositories;
+import repositories.SQLConnection;
+import repositories.UserTable;
 
 import java.sql.SQLException;
 import java.util.Scanner;
 
 public class UserController {
-    private Repositories userRepository;
+    private UserTable userRepository;
 
-    public UserController(Repositories userRepository) {
+    public UserController(UserTable userRepository) {
         this.userRepository = userRepository;
     }
 
@@ -22,12 +23,12 @@ public class UserController {
         System.out.print("Enter balance: ");
         double balance = Double.parseDouble(scanner.nextLine());
 
-        userRepository.addUser(new User(username, age, balance));
+        userRepository.addToTable(new User(username, age, balance));
     }
     public void updateUser(Scanner scanner) throws SQLException{
             System.out.println("Enter the id to update");
             int id = Integer.parseInt(scanner.nextLine());
-            if(userRepository.getUser(id)) {
+            if(userRepository.getObjectFromTable(id)) {
                 System.out.println("Enter the surname to update");
                 String username = scanner.nextLine();
 
@@ -37,25 +38,25 @@ public class UserController {
                 System.out.println("Enter the balance to update");
                 double balance = Double.parseDouble(scanner.nextLine());
 
-                userRepository.updateUser(new User(id, username, age, balance));
+                userRepository.updateObjectTable(new User(id, username, age, balance));
             }
     }
     public void deleteUser(Scanner scanner) throws SQLException{
                 System.out.println("Enter the id to delete");
                 int id = Integer.parseInt(scanner.nextLine());
-                if(userRepository.getUser(id)) {
-                    userRepository.deleteUser(id);
+                if(userRepository.getObjectFromTable(id)) {
+                    userRepository.deleteObjectFromTable(id);
                 }
     }
      public void getUser(Scanner scanner) throws SQLException{
         Scanner sc = new Scanner(System.in);
         System.out.println("Enter id of username");
         int id = sc.nextInt();
-        userRepository.getUser(id);
+        userRepository.getObjectFromTable(id);
      }
 
 
     public void getAllUsers() throws SQLException{
-        userRepository.getAllUsers();
+        userRepository.getAllObjectsFromTable();
     }
 }
